Part of a massive change to update the core of the library.
This is incomplete.

This takes many features from the new_ab branch like vectors, hashtables etc.
and adds them onto the older 2.0 branch.   The idea is to keep the packet handling
roughly the same rather than using the new methods.  Jobs are also added.  We will
keep the hashtable lookup for tags.

This refactors the tag vtable a bit to add four new functions for getting and setting
integers (int64) and floats (double).  Each protocol is responsible for its own implementation
of these functions.

This also starts the refactoring of the platform code into separate pieces in the
util directory.  The goal is to move all platform dependent code into #ifdef guarded
sections and remove all platform dependent file selection in the build system.