In process of changing over to PLC-first creation in ab_tag_create().
This will reorder the logic in the tag creation completely and push more of
the pieces over to the session/plc object.

Possible refactor: change the tag creation part completely to just be done
within the PLC object.  Split out all the PLC types from each other.   This would result
in a lot of duplicate code, but could simplify a lot of things.   Refactoring after that
would help pull in the bloat.